// src/app/models/students.ts
export interface Students {
    $uid?: string;
    $key?: string;
    firstName?: string;
    lastName?: string;
    email?: string;
    password?: string;
    phone?: string;
    university? :string;
    educationStatus?: string[];
    courseType?: string[];
    major?: string[];
    year?: string[];
    semester?: string[];
    skill_1?: string;
    skill_2?: string;
    skill_3?: string;
    skill_4?: string;
    lookingProject?: string[];
    projectType?: string[];
    visaStatus?: string[];
    coverLetter?: string;
}